+++ libs/cardano-ledger-core/testlib/Test/Cardano/Ledger/Imp/Common.hs
@@ -73,6 +73,8 @@ module Test.Cardano.Ledger.Imp.Common (
 where
 
 import Control.Monad.IO.Class
+import Data.Functor.Const (Const (..))
+import Data.Functor.Identity (Identity (..))
 import Test.Cardano.Ledger.Binary.TreeDiff (expectExprEqualWithMessage)
 import Test.Cardano.Ledger.Common as X hiding (
   arbitrary,
@@ -354,7 +356,18 @@ instance
 class HasSubState s where
   type SubState s :: Type
   getSubState :: s -> SubState s
+  getSubState = getConst . subStateL Const
   setSubState :: s -> SubState s -> s
+  setSubState s a = runIdentity $ subStateL (const $ Identity a) s
+  subStateL :: Functor f => (SubState s -> f (SubState s)) -> (s -> f s)
+  subStateL k s = setSubState s <$> k (getSubState s)
+  {-# MINIMAL subStateL | getSubState, setSubState #-}
+
+-- | Modify the sub-state and return a value, using the supplied function.
+-- Similar to the `state` method of `MonadState`.
+subState :: (HasSubState s, MonadState s m) => (SubState s -> (a, SubState s)) -> m a
+subState = state . subStateL -- Uses (a,) as the functor for subStateL
+{-# INLINE subState #-}
 
 uniformM ::
   ( HasStatefulGen g m
@@ -411,12 +424,6 @@ instance HasSubState (StateGen g) where
   {-# INLINE getSubState #-}
   setSubState _ = StateGen
   {-# INLINE setSubState #-}
-
-subState :: (HasSubState s, MonadState s m) => (SubState s -> (a, SubState s)) -> m a
-subState f = state $ \s ->
-  case f (getSubState s) of
-    (a, g) -> (a, setSubState s g)
-{-# INLINE subState #-}
 
 instance
   (HasSubState s, R.RandomGen (SubState s), MonadState s m) =>